[FEAT] #20 - Alert extension으로 구분
