+++ NADA-iOS-forRelease/Sources/ViewControllers/CardList/CardListViewController.swift
@@ -81,30 +81,3 @@ extension CardListViewController: UITableViewDataSource {
         return serviceCell
     }
 }
-
-// Alert창 구현
-extension CardListViewController {
-    func makeAlert(title: String,
-                   message: String,
-                   cancelAction: ((UIAlertAction) -> Void)? = nil,
-                   deleteAction: ((UIAlertAction) -> Void)?,
-                   completion: (() -> Void)? = nil) {
-        let genetator = UIImpactFeedbackGenerator(style: .medium)
-        genetator.impactOccurred()
-        
-        let alertViewController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        
-        alertViewController.setTitlet(font: UIFont.boldSystemFont(ofSize: 17), color: UIColor.white)
-        alertViewController.setMessage(font: UIFont.systemFont(ofSize: 13), color: UIColor.white)
-        
-        alertViewController.view.subviews.first?.subviews.first?.subviews.first?.backgroundColor = UIColor(red: 30/255, green: 30/255, blue: 30/255, alpha: 3/4)
-        
-        let cancelAction = UIAlertAction(title: "취소", style: .cancel, handler: cancelAction)
-        alertViewController.addAction(cancelAction)
-        
-        let deleteAction = UIAlertAction(title: "삭제", style: .default, handler: deleteAction)
-        alertViewController.addAction(deleteAction)
-        
-        self.present(alertViewController, animated: true, completion: completion)
-    }
-}
